refactor: modularize renderer plugin runtime components
